# v2/cmd/postmangen/testdata/openapi.golden.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Collection (Generated)",
    "description": "Generated from registered Go routes and request structs.",
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this page"
    },
    {
      "url": "http://localhost:3000",
      "description": "Configured base URL"
    }
  ],
...
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "GET /users/{id}",
        "description": "Handled by `UserController.Find`.",
        "operationId": "UserControllerFind",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "object",
                      "properties": {
                        "district": {
                          "type": "object",
                          "properties": {
                            "addresses": {
                              "type": "array",
                              "items": {
                                "type": "object"
                              }
                            },
                            "id": {
                              "type": "string"
                            },
                            "name_th": {
                              "type": "string"
                            },
                            "province": {
                              "type": "object",
                              "properties": {
                                "districts": {
                                  "type": "array",
                                  "items": {
                                    "type": "object"
                                  }
                                },
                                "id": {
                                  "type": "string"
                                },
                                "name_en": {
                                  "type": "string"
                                },
                                "name_th": {
                                  "type": "string"
                                }
                              }
                            },
                            "province_id": {
                              "type": "string"
                            }
                          }
                        },
                        "district_id": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "line": {
                          "type": "string"
                        },
                        "province": {
                          "type": "object",
                          "properties": {
                            "districts": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "addresses": {
                                    "type": "array",
                                    "items": {
                                      "type": "object"
                                    }
                                  },
                                  "id": {
                                    "type": "string"
                                  },
                                  "name_th": {
                                    "type": "string"
                                  },
                                  "province": {
                                    "type": "object"
                                  },
                                  "province_id": {
                                    "type": "string"
                                  }
                                }
                              }
                            },
                            "id": {
                              "type": "string"
                            },
                            "name_en": {
                              "type": "string"
                            },
                            "name_th": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "email": {
                      "type": "string"
                    },
                    "full_name": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "address": {
                    "district": {
                      "addresses": [
                        {}
                      ],
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "name_th": "Sample name",
                      "province": {
                        "districts": [
                          {}
                        ],
                        "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "name_en": "Sample name",
                        "name_th": "Sample name"
                      },
                      "province_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "district_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "line": "string",
                    "province": {
                      "districts": [
                        {
                          "addresses": [
                            {}
                          ],
                          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                          "name_th": "Sample name",
                          "province": {},
                          "province_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        }
                      ],
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "name_en": "Sample name",
                      "name_th": "Sample name"
                    }
                  },
                  "email": "user@example.com",
                  "full_name": "Jane Doe",
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "unauthorized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}