# v2/http_context.go
package core

import (
	"bytes"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"mime/multipart"
	"net/http"
	"net/url"
	"reflect"
	"runtime"
	"strconv"
	"strings"

	"github.com/labstack/echo/v5"
)

// IHTTPContext is the per-request HTTP context: it is an IContext (all
// capabilities) plus the request/response surface of echo. Same name as v1.
//
// echo v5 turned Context from an interface into a struct, so this interface can
// no longer embed it. The methods handlers actually use are declared here
// instead and satisfied by the embedded *echo.Context — call sites are
// unchanged (`c.JSON`, `c.Param`, `c.Bind` … still work), and anything not
// listed is reachable through Echo().
type IHTTPContext interface {
	IContext

	// --- request ---
	Request() *http.Request
	SetRequest(r *http.Request)
	// Response is the raw writer. echo's own recorder (Status, Size, Committed)
	// is reached with echo.UnwrapResponse.
	Response() http.ResponseWriter
	IsTLS() bool
	IsWebSocket() bool
	Scheme() string
	RealIP() string

	// --- routing ---
	// Path is the registered route pattern ("/users/:id"), not the concrete URL.
	Path() string
	RouteInfo() echo.RouteInfo
	Param(name string) string
	ParamOr(name, defaultValue string) string
	PathValues() echo.PathValues
	QueryParam(name string) string
	QueryParamOr(name, defaultValue string) string
	QueryParams() url.Values
	QueryString() string

	// --- payload ---
	FormValue(name string) string
	FormValueOr(name, defaultValue string) string
	FormValues() (url.Values, error)
	FormFile(name string) (*multipart.FileHeader, error)
	MultipartForm() (*multipart.Form, error)
	Cookie(name string) (*http.Cookie, error)
	Cookies() []*http.Cookie
	SetCookie(cookie *http.Cookie)
	Bind(i any) error
	Validate(i any) error

	// --- per-request values (echo's own store, not ctx.SetData) ---
	Get(key string) any
	Set(key string, val any)

	// --- responses ---
	JSON(code int, i any) error
	JSONPretty(code int, i any, indent string) error
	JSONBlob(code int, b []byte) error
	XML(code int, i any) error
	String(code int, s string) error
	HTML(code int, html string) error
	Blob(code int, contentType string, b []byte) error
	Stream(code int, contentType string, r io.Reader) error
	File(file string) error
	Attachment(file, name string) error
	Inline(file, name string) error
	Render(code int, name string, data any) error
	NoContent(code int) error
	Redirect(code int, url string) error

	// Echo is the escape hatch to the underlying instance.
	Echo() *echo.Echo

	BindWithValidate(payload IValidateContext) IError
	BindOnly(payload any) IError
	GetPageOptions() *PageOptions
	GetPageOptionsWithAllowed(allowed ...string) *PageOptions
}

// HandlerFunc is an mine-core HTTP handler.
type HandlerFunc func(c IHTTPContext) error

type httpContext struct {
	IContext
	*echo.Context
}

var _ IHTTPContext = (*httpContext)(nil)

// WithHTTPContext adapts an mine-core handler to echo, constructing the
// per-request context from the App (pools are not recreated per request). If an
// auth middleware has resolved a user, it is attached so c.GetUser() works.
func WithHTTPContext(app *App, h HandlerFunc) echo.HandlerFunc {
	// resolved once, at registration: by the time the access line is written the
	// handler has returned and its frame is gone from the stack
	name := handlerName(h)
	return func(ec *echo.Context) error {
		if name != "" {
			ec.Set(handlerContextKey, name)
		}
		// every middleware has run by now, so the body limit in force is final:
		// this is where a body that declares an oversized length is refused
		// before the handler is entered, whether or not the handler reads it
		if err := checkDeclaredBodySize(ec); err != nil {
			return err
		}

		ctx := app.NewContext(ec.Request().Context(), ModeHTTP)
		if user, ok := ec.Get(userContextKey).(*ContextUser); ok && user != nil {
			ctx.SetUser(user)
		}
		bindSentryScope(ec.Request().Context(), ctx, ec.RealIP())
		return h(&httpContext{IContext: ctx, Context: ec})
	}
}

// handlerContextKey is where WithHTTPContext leaves the handler's name for the
// access line, which runs after the handler has returned.
const handlerContextKey = "core.handler"

// handlerName is the name of the function serving a route, as the access line
// reports it: "user.(*Controller).Login".
//
// A name and not a file and line, which is what every other log line carries.
// The usual way to register a route is a method value (`ctrl.Login`), and Go
// compiles one into a generated wrapper whose source position is
// "<autogenerated>:1" — there is no file to report. The name survives that: the
// wrapper is the method's own, with "-fm" appended.
func handlerName(h HandlerFunc) string {
	if h == nil {
		return ""
	}
	fn := runtime.FuncForPC(reflect.ValueOf(h).Pointer())
	if fn == nil {
		return ""
	}
	name := strings.TrimSuffix(fn.Name(), "-fm")
	// the import path in front is the same for every route in the service and
	// would push the part that differs off the end of the line
	if i := strings.LastIndexByte(name, '/'); i >= 0 {
		name = name[i+1:]
	}
	return name
}

// BindOnly binds every request source into payload — path params (`param:` tag),
// query params (`query:` tag), and the body (JSON/XML/form/multipart, chosen by
// Content-Type; `json:`/`xml:`/`form:` tags) — like echo, without validating.
// Uploaded files are read via c.FormFile / c.MultipartForm.
func (c *httpContext) BindOnly(payload any) IError {
	if err := echo.BindPathValues(c.Context, payload); err != nil {
		return bindError(err)
	}
	if err := echo.BindQueryParams(c.Context, payload); err != nil {
		return bindError(err)
	}
	return c.bindBody(payload)
}

// bindBody binds the request body. JSON is handled directly so type mismatches
// produce a field-level error; other content types delegate to echo's binder.
func (c *httpContext) bindBody(payload any) IError {
	if c.Request().Body == nil || c.Request().ContentLength == 0 {
		return nil
	}
	ctype := c.Request().Header.Get(echo.HeaderContentType)
	if strings.HasPrefix(ctype, echo.MIMEApplicationJSON) {
		body, err := c.readBody()
		if err != nil {
			return err
		}
		if uerr := json.Unmarshal(body, payload); uerr != nil {
			return unmarshalError(uerr)
		}
		return nil
	}
	// form-urlencoded, multipart/form-data, XML → echo's binder
	if err := echo.BindBody(c.Context, payload); err != nil {
		return bindError(err)
	}
	return nil
}

// BindWithValidate binds all request sources then runs payload.Valid(ctx).
func (c *httpContext) BindWithValidate(payload IValidateContext) IError {
	if err := c.BindOnly(payload); err != nil {
		return err
	}
	err := payload.Valid(c)
	if err != nil {
		// label each violation with the part of the request it came from — the
		// payload's binding tags already say, and only this layer can see both
		markFieldSources(payload, err)
	}
	return err
}

// bindError maps an echo bind failure to a 400 IError. The cause is kept: a body
// cut off by the size limit fails the binder as malformed input, and only the
// cause says it was actually too large.
func bindError(err error) IError {
	var he *echo.HTTPError
	if errors.As(err, &he) {
		return New(he.Code, "INVALID_PARAMS", he.Message).WithCause(err)
	}
	return New(http.StatusBadRequest, "INVALID_PARAMS", err.Error()).WithCause(err)
}

func (c *httpContext) readBody() ([]byte, IError) {
	if c.Request().Body == nil {
		return []byte("{}"), nil
	}
	body, err := io.ReadAll(c.Request().Body)
	if err != nil {
		// keep the cause: a body cut off by the size limit fails here, and only
		// the cause says so — BodyLimit re-reads it to answer 413 instead of a
		// 400 that blames the client's JSON
		return nil, New(http.StatusBadRequest, "INVALID_BODY", "cannot read request body").
			WithCause(err)
	}
	// restore for any downstream reader
	c.Request().Body = io.NopCloser(bytes.NewBuffer(body))
	if len(body) == 0 {
		return []byte("{}"), nil
	}
	return body, nil
}

func unmarshalError(err error) IError {
	var typeErr *json.UnmarshalTypeError
	if e, ok := err.(*json.UnmarshalTypeError); ok {
		typeErr = e
	}
	if typeErr != nil {
		return New(http.StatusBadRequest, "INVALID_PARAMS", "invalid parameters").
			WithFields(map[string]any{
				typeErr.Field: map[string]string{
					"code":    "INVALID_TYPE",
					"message": fmt.Sprintf("The %s field must be %s", typeErr.Field, typeErr.Type.String()),
				},
			})
	}
	return New(http.StatusBadRequest, "INVALID_JSON", "Must be json format")
}

// GetPageOptions parses limit/page/q/order_by from the query string.
//
// Any order_by column that is not a plain (optionally table-qualified)
// identifier is dropped, so nothing a client sends can become SQL. That is the
// floor, not the whole story: it still lets a caller sort by any real column,
// including ones the endpoint never meant to expose. Name the columns you mean
// with GetPageOptionsWithAllowed.
func (c *httpContext) GetPageOptions() *PageOptions {
	limit, _ := strconv.ParseInt(c.QueryParam("limit"), 10, 64)
	page, _ := strconv.ParseInt(c.QueryParam("page"), 10, 64)
	opts := &PageOptions{
		Q:       c.QueryParam("q"),
		Limit:   limit,
		Page:    page,
		OrderBy: parseOrderBy(c.QueryParam("order_by"), nil),
	}
	opts.normalize()
	return opts
}

// GetPageOptionsWithAllowed is GetPageOptions restricted to an order-by allowlist
// (protects against SQL injection through order_by).
func (c *httpContext) GetPageOptionsWithAllowed(allowed ...string) *PageOptions {
	opts := c.GetPageOptions()
	opts.OrderBy = parseOrderBy(c.QueryParam("order_by"), allowed)
	return opts
}

// parseOrderBy and its column check live in pagination.go, next to PageOptions.